Move enzyme-specific serialization to subpackage
