+++ packages/jest-emotion/enzyme/index.js
@@ -1,0 +1,1 @@
+export * from '../src/enzyme'
