+++ packages/jest-emotion/enzyme/package.json
@@ -1,0 +1,8 @@
+{
+  "main": "dist/jest-emotion.cjs.js",
+  "module": "dist/jest-emotion.esm.js",
+  "types": "../types/enzyme",
+  "preconstruct": {
+    "source": "../src/enzyme"
+  }
+}
