+++ packages/jest-emotion/package.json
@@ -3,11 +3,13 @@
   "version": "11.0.0-next.7",
   "description": "Jest utilities for emotion",
   "main": "dist/jest-emotion.cjs.js",
+  "module": "dist/jest-emotion.esm.js",
   "types": "types/index.d.ts",
   "files": [
     "src",
     "dist",
     "types/*.d.ts",
+    "enzyme",
     "serializer.js"
   ],
   "scripts": {
@@ -19,6 +21,9 @@
     "chalk": "^2.4.1",
     "css": "^2.2.1",
     "specificity": "^0.4.1"
+  },
+  "peerDependencies": {
+    "enzyme-to-json": "^3.2.1"
   },
   "devDependencies": {
     "@emotion/core": "^11.0.0-next.7",
@@ -47,7 +52,10 @@
   "bugs": {
     "url": "https://github.com/emotion-js/emotion/issues"
   },
-  "browser": {
-    "./dist/jest-emotion.cjs.js": "./dist/jest-emotion.browser.cjs.js"
+  "preconstruct": {
+    "entrypoints": [
+      ".",
+      "enzyme"
+    ]
   }
 }
