+++ packages/jest-emotion/types/enzyme.d.ts
@@ -1,0 +1,1 @@
+/// <reference types="jest" />
