+++ packages/styled/test/theming.dom.test.js
@@ -4,9 +4,6 @@ import React from 'react'
 import styled from '@emotion/styled'
 import { ThemeProvider } from '@emotion/core'
 import { render, cleanup } from '@testing-library/react'
-import serializer from 'jest-emotion'
-
-expect.addSnapshotSerializer(serializer)
 
 afterEach(cleanup)
