+++ scripts/test-utils/next-env.js
@@ -1,5 +1,5 @@
 // @flow
 /* eslint-env jest */
-import serializer from 'jest-emotion'
+import serializer from 'jest-emotion/enzyme'
 
 expect.addSnapshotSerializer(serializer)
